Rename old Phrases folder and adjust relevant imports
