+++ frontend/app/common/conf/pagesIndex.js
@@ -38,7 +38,7 @@ import PageDialectReports from 'components/Reports'
 import PageDialectReportsView from 'components/Reports/ReportsView'
 
 import PageDialectLearnWords from 'components/Words/WordsContainer'
-import PageDialectLearnPhrases from 'components/Phrases'
+import PageDialectLearnPhrases from 'components/PhrasesCreateEdit'
 import PageDialectLearnStoriesAndSongs from 'components/SongsStories'
 
 import PageDialectViewMedia from 'components/Media/view'
@@ -51,7 +51,7 @@ import PageDialectLearnWordsCategories from 'components/Categories/WordCategorie
 
 import PhraseBooksGrid from 'components/PhraseBooksGrid'
 import WordsCategoriesGrid from 'components/WordsCategoriesGrid'
-import PageDialectLearnPhrasesByPhrasebook from 'components/Phrases/phrasesFilteredByPhrasebook'
+import PageDialectLearnPhrasesByPhrasebook from 'components/PhrasesCreateEdit/phrasesFilteredByPhrasebook'
 
 import PageDialectImmersionList from 'components/Immersion'
 
@@ -78,7 +78,7 @@ import PageExploreDialectEdit from 'components/ExploreDialectEdit'
 import PageDialectGalleryEdit from 'components/Gallery/edit'
 import PageDialectEditMedia from 'components/Media/edit'
 import PageDialectWordEdit from 'components/WordsCreateEdit/Edit'
-import PageDialectPhraseEdit from 'components/Phrases/Edit'
+import PageDialectPhraseEdit from 'components/PhrasesCreateEdit/Edit'
 import PageDialectBookEdit from 'components/SongsStories/edit'
 import PageDialectBookEntryEdit from 'components/SongsStories/entry/edit'
 import PageDialectAlphabetCharacterEdit from 'components/Alphabet/edit'
@@ -87,7 +87,7 @@ import PageDialectAlphabetCharacterEdit from 'components/Alphabet/edit'
 import { default as PageDialectWordsCreate } from 'components/WordsCreateEdit/Create'
 import { default as CreateV2 } from 'components/WordsCreateEdit/CreateV2'
 import { default as CreateAudio } from 'components/Audio'
-import { default as PageDialectPhrasesCreate } from 'components/Phrases/Create'
+import { default as PageDialectPhrasesCreate } from 'components/PhrasesCreateEdit/Create'
 import { default as PageDialectStoriesAndSongsCreate } from 'components/SongsStories/create'
 import { default as PageDialectStoriesAndSongsBookEntryCreate } from 'components/SongsStories/entry/create'
 import { default as PageDialectGalleryCreate } from 'components/Gallery/create'
