+++ frontend/app/components/AlphabetDetail/AlphabetDetailPresentation.js
@@ -27,7 +27,7 @@ import FVTab from 'components/FVTab'
 import LiteralUnicodeData from 'components/LiteralUnicode/LiteralUnicodeData'
 import LiteralUnicodePresentation from 'components/LiteralUnicode/LiteralUnicodePresentation'
 import PageToolbar from 'components/PageToolbar'
-import PhraseListView from 'components/Phrases/list-view'
+import PhraseListView from 'components/PhrasesCreateEdit/list-view'
 import Preview from 'components/Preview'
 import PromiseWrapper from 'components/PromiseWrapper'
 import WordListView from 'components/WordsCreateEdit/list-view'
