+++ frontend/app/components/PhrasesCreateEdit/list-view.js
